+++ packages/wezterm/pink.lua
@@ -1,0 +1,67 @@
+return {
+  foreground = "#fff8f8",
+  background = "#202330",
+  cursor_bg = "#fff8f8",
+  cursor_fg = "#202330",
+  cursor_border = "#fff8f8",
+  selection_fg = "#fff8f8",
+  selection_bg = "#472541",
+  scrollbar_thumb = "#6d5d72",
+  split = "#6d5d72",
+  ansi = {
+    "#2d2f42",
+    "#FA508C",
+    "#58B896",
+    "#d0963a",
+    "#DCBFF2",
+    "#ff62a5",
+    "#9CD162",
+    "#fff8f8"
+  },
+  brights = {
+    "#6d5d72",
+    "#FF7F9D",
+    "#37E884",
+    "#FAE8B6",
+    "#FFC85B",
+    "#E6A1FF",
+    "#A2C2EB",
+    "#ffffff"
+  },
+  indexed = {
+    [16] = "#FFC85B",
+    [17] = "#F5A6C6"
+  },
+  compose_cursor = "#FA508C",
+  copy_mode_active_highlight_bg = { Color = "#FA508C" },
+  copy_mode_active_highlight_fg = { Color = "#202330" },
+  copy_mode_inactive_highlight_bg = { Color = "#472541" },
+  copy_mode_inactive_highlight_fg = { Color = "#fff8f8" },
+  quick_select_label_bg = { Color = "#FA508C" },
+  quick_select_label_fg = { Color = "#202330" },
+  quick_select_match_bg = { Color = "#FAE8B6" },
+  quick_select_match_fg = { Color = "#202330" },
+  tab_bar = {
+    background = "#202330",
+    active_tab = {
+      bg_color = "#FA508C",
+      fg_color = "#202330"
+    },
+    inactive_tab = {
+      bg_color = "#2d2f42",
+      fg_color = "#F5A6C6"
+    },
+    inactive_tab_hover = {
+      bg_color = "#472541",
+      fg_color = "#fff8f8"
+    },
+    new_tab = {
+      bg_color = "#202330",
+      fg_color = "#DCBFF2"
+    },
+    new_tab_hover = {
+      bg_color = "#472541",
+      fg_color = "#fff8f8"
+    }
+  }
+}